+++ index.php
@@ -15,8 +15,10 @@ declare(strict_types=1);
 
 namespace axy\errors;
 
+use LogicException;
+
 if (!is_file(__DIR__.'/vendor/autoload.php')) {
-    throw new \LogicException('Please: composer install');
+    throw new LogicException('Please: composer install');
 }
 
 require_once(__DIR__.'/vendor/autoload.php');